+++ src/main/java/Grammar/Expression/BinaryArithExpr.java
@@ -7,9 +7,9 @@ import java.util.Map;
 
 public record BinaryArithExpr(Expression left, String op, Expression right) implements Expression {
     @Override
-    public double eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError {
-        double lv = left.eval(bindings,p,t);
-        double rv = right.eval(bindings,p,t);
+    public long eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError {
+        long lv = left.eval(bindings,p,t);
+        long rv = right.eval(bindings,p,t);
         switch (op) {
             case "+" -> {
                 return lv + rv;
@@ -27,7 +27,7 @@ public record BinaryArithExpr(Expression left, String op, Expression right) impl
                 return lv % rv;
             }
             case "^" -> {
-                return (int) Math.pow(lv, rv);
+                return (long) Math.pow(lv, rv);
             }
         }
         throw new EvalError("unknown op: " + op);
